Remove redundant token type assignments.
Previously, we had a lot of redundant assignments upon the return from a sub-
function.  Now we save some copies by assigning the token type at the beginn-
ing of the dispatch.  When later in the process the token type changes, it has
to be assigned manually.
the token types